api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arch==8.0.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
fastapi==0.128.0
finfetcher==0.1.0
fonttools==4.61.1
frozendict==2.4.7
h11==0.16.0
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
multitasking==0.0.12
narwhals==2.15.0
numpy==2.4.1
pandas==3.0.0
patsy==1.0.2
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
protobuf==6.33.4
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.50.0
statsmodels==0.14.6
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==16.0
wheel==0.46.3
